requirements.txt
torch 
ultralytics 
opencv-python 
pillow 
numpy